Import consolidation
