+++ scripts/app/CustomCompendiums.js
@@ -1,5 +1,4 @@
-import { HM } from '../hero-mancer.js';
-import { CacheManager } from '../utils/index.js';
+import { CacheManager, HM } from '../utils/index.js';
 
 const { ApplicationV2, HandlebarsApplicationMixin, DialogV2 } = foundry.applications.api;
